Move tensor contraction attributes to new folder
