+++ src/Mathematics.NET/Core/Attributes/GenerateTensorSelfContractions.cs
@@ -25,7 +25,7 @@
 // SOFTWARE.
 // </copyright>
 
-namespace Mathematics.NET.Core.Attributes.GeneratorAttributes;
+namespace Mathematics.NET.Core.Attributes;
 
 /// <summary>Indicates that index permutations of a tensor self-contraction should be generated.</summary>
 [AttributeUsage(AttributeTargets.Method, AllowMultiple = false)]
